% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalUni-class.R
\docType{class}
\name{animalUni-class}
\alias{animalUni-class}
\alias{amimalUni}
\title{AnimalUni that extends \code{"stanfit"} class.}
\description{
AnimalUni provides some resonable print and summary methods for animal model fits
}
