% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateAnimalModelInput.R
\name{genAnimalModelInput}
\alias{genAnimalModelInput}
\title{Generate data input for an animal model.}
\usage{
genAnimalModelInput(
  formula,
  data,
  A,
  IDcol = "ID",
  out_type = c("none", "HDF5", "matlab"),
  out_folder = "BSFG_run"
)
}
\arguments{
\item{formula}{formula specifying fixed effects and response variables}

\item{data}{data.frame with data.}

\item{A}{relationship matrix from pedigree}

\item{IDcol}{Which column has the individual IDs which match the A matrix rownames.}

\item{out_type}{Default is none to output object to R. Choose 'HDF5' or 'matlab' for mat}

\item{out_folder}{folder to create the Matlab or Julia input files}
}
\value{
list with animal model input objects
}
\description{
Generate data input for an animal model.
}
\examples{
library(ratones)
data(ratonesdf)
data(ratones_ped)
formula = paste0("cbind(",
          paste(names(dplyr::select(ratonesdf, IS_PM:BA_OPI)),
                collapse = ", "), ") ~ SEX + AGE")
animal_model_data = generateAnimalModelInput(formula, ratonesdf,
                                             ratones_ped$A, out_folder = "./")
}
