% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_multiCov.R
\name{lmm_multiCov}
\alias{lmm_multiCov}
\title{Fit a bayesian covariance matrix}
\usage{
lmm_multiCov(Y, X, LKJ = 4, iter = 3000, warmup = 2000, chains = 4,
  control = list(adapt_delta = 0.99), ...)
}
\arguments{
\item{Y}{response variable, univariate or multivariate}

\item{X}{fixed model matrix}

\item{LKJ}{parameter of the LKJ prior on the correlations. Higher values result in lower correlations. A LKJ=1 implies uniform prior on the correlations. Default is 4, which tends to give reasonable amount of shrinkage.}

\item{iter}{number of MCMC iterations}

\item{warmup}{number of warmup iterations, must be smaller than iter}

\item{chains}{number of chains to run}

\item{control}{list of stan control arguments}

\item{...}{additional parameters for Stan sampling}
}
\description{
Fits a multivariate covariance matrix for the residuals a linear model.
}
\author{
Diogo Melo
}
