% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateAnimalModelInput.R
\name{saveHDF5animalModel}
\alias{saveHDF5animalModel}
\title{Generate input for BSFG model in Julia}
\usage{
saveHDF5animalModel(Y, X, A, Z, out_file)
}
\arguments{
\item{Y}{matrix of observed phenotypes}

\item{X}{fixed effects model matrix}

\item{A}{relationship matrix from pedigree. Use nadiv package.}

\item{Z}{Which column has the individual IDs which match the A matrix rownames.}

\item{out_file}{output filename.}
}
\value{
list with animal model input objects
}
\description{
Generate input for BSFG model in Julia
}
\examples{
library(ratones)
saveHDF5animalModel(Y, X, A, Z, "setup.h5")
}
