% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanAnimal-package.R
\docType{package}
\name{stanAnimal-package}
\alias{stanAnimal-package}
\alias{package2}
\title{The 'stanAnimal' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
