% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_animal.R
\name{lmm_animal}
\alias{lmm_animal}
\title{Fit linear mixed animal model}
\usage{
lmm_animal(
  Y,
  X,
  A,
  iter = 3000,
  warmup = floor(iter/2),
  chains = 4,
  control = list(adapt_delta = 0.99),
  ...
)
}
\arguments{
\item{Y}{response variable, univariate or multivariate}

\item{X}{fixed model matrix}

\item{A}{relatedness matrix}

\item{iter}{number of MCMC iterations}

\item{warmup}{number of warmup iterations, must be smaller than iter}

\item{chains}{number of chains to run}

\item{control}{list of stan control arguments}

\item{...}{additional parameters for stan sampling}
}
\description{
Fits a univariate or multivariate linear mixed animal model using a know relatedness matrix
}
\author{
Diogo Melo
}
