% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanAnimal-package.R
\docType{package}
\name{stanAnimal-package}
\alias{stanAnimal-package}
\alias{stanAnimal}
\title{The 'stanAnimal' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
